% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucketing.R
\name{append_bucket_column}
\alias{append_bucket_column}
\title{Append a Bucketed Column to a Database Table Based on a Lookup Table}
\usage{
append_bucket_column(con, filepath, data, input_column, id_column)
}
\arguments{
\item{con}{A database connection.}

\item{filepath}{The path to the lookup table CSV file.}

\item{data}{A lazy database reference to the input data.}

\item{input_column}{The column to be bucketed/matched.}

\item{id_column}{The unique ID column in the data.}
}
\value{
A lazy database reference with the appended bucketed column.
}
\description{
This function adds a new bucketed column to a database table using a range-based or value-based lookup table.
The lookup table is provided as a CSV file, and the function splits the lookup table into value and range components.
}
