% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{crosstab_count_no_se}
\alias{crosstab_count_no_se}
\title{Calculate Weighted and Unweighted Counts for Groups (No Standard Error)}
\usage{
crosstab_count_no_se(data, wt, group_by, every_combo = FALSE)
}
\arguments{
\item{data}{A data frame or a database connection object containing the data to be aggregated.}

\item{wt}{A string specifying the name of the column containing the weights.}

\item{group_by}{A character vector of column names to group by.}

\item{every_combo}{Logical, whether to include all combinations of the grouping variables,
setting counts to zero for combinations not present in the data. Defaults to \code{FALSE}.}
}
\value{
A tibble or database connection object containing the group-by columns, weighted count,
and unweighted count for each group.
}
\description{
This function calculates the weighted count (sum of weights) and the unweighted count
of observations for the specified weight column, grouped by the specified columns.
Optionally, it can include all combinations of the grouping variables, even if some combinations
do not exist in the data, setting the counts to zero for those combinations.
}
