% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{crosstab_mean}
\alias{crosstab_mean}
\title{Calculate Weighted Mean for Groups}
\usage{
crosstab_mean(
  data,
  value,
  weight,
  group_by,
  every_combo = FALSE,
  repwts = paste0("REPWTP", sprintf("\%d", 1:80))
)
}
\arguments{
\item{data}{A data frame or a database connection object containing the data to be aggregated.}

\item{value}{A string specifying the name of the column containing the values to be averaged.}

\item{weight}{A string specifying the name of the column containing the weights.}

\item{group_by}{A character vector of column names to group by.}
}
\value{
A tibble or database connection object containing the group-by columns and weighted mean for each group.
}
\description{
This function calculates the weighted mean for the specified value and weight columns, grouped by the specified columns.
}
