% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{crosstab_mean_no_se}
\alias{crosstab_mean_no_se}
\title{Calculate Weighted Mean for Groups (No Standard Error)}
\usage{
crosstab_mean_no_se(data, value, wt, group_by, every_combo = FALSE)
}
\arguments{
\item{data}{A data frame or a database connection object containing the data to be aggregated.}

\item{value}{A string specifying the name of the column containing the values to be averaged.}

\item{wt}{A string specifying the name of the column containing the weights.}

\item{group_by}{A character vector of column names to group by.}

\item{every_combo}{Logical, whether to include all combinations of the grouping variables,
setting counts to zero for combinations not present in the data. Defaults to \code{FALSE}.}
}
\value{
A tibble or database connection object containing the group-by columns and weighted mean for each group.
}
\description{
This function calculates the weighted mean for the specified value and weight columns, grouped by the specified columns.
Optionally, it can include all combinations of the grouping variables, even if some combinations
do not exist in the data, setting the counts to zero for those combinations.
}
