% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{crosstab_percent_no_se}
\alias{crosstab_percent_no_se}
\title{Calculate Percentages Within Groups (No Standard Error)}
\usage{
crosstab_percent_no_se(
  data,
  wt,
  group_by,
  percent_group_by,
  every_combo = FALSE
)
}
\arguments{
\item{data}{A data frame or a database connection object containing the data to be aggregated.}

\item{group_by}{A character vector of column names to group by for the main counts.}

\item{percent_group_by}{A character vector of column names to group by for the percentage calculation.
All elements of \code{percent_group_by} must be included in \code{group_by}.}

\item{every_combo}{Logical, whether to include all combinations of the grouping variables,
setting counts to zero for combinations not present in the data. Defaults to \code{FALSE}.}

\item{weight}{A string specifying the name of the column containing the weights.}
}
\value{
A tibble or database connection object containing the group-by columns and percentages for each group.
}
\description{
This function calculates percentages of weighted counts within specified groups.
Optionally, it can include all combinations of the grouping variables, even if some combinations
do not exist in the data, setting the counts to zero for those combinations.
}
