% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{difference_means}
\alias{difference_means}
\title{Calculate the Difference in Means Between Two Datasets}
\usage{
difference_means(data2000, data2020, match_by, diff_by, keep = NULL)
}
\arguments{
\item{data2000}{A data frame or database object representing the first dataset, such as data from the year 2000.}

\item{data2020}{A data frame or database object representing the second dataset, such as data from the year 2020.}

\item{match_by}{A string specifying the column name used to match the two datasets.}

\item{diff_by}{A string specifying the column along which to calculate differences (e.g., a measure like income or household size).}

\item{keep}{A character vector of additional column names to keep in the output, along with the difference column. Defaults to NULL.}
}
\value{
A tibble containing the matching column, the calculated difference for \code{diff_by},
and any additional columns specified in \code{keep}.
}
\description{
This function calculates the difference in means between two datasets (such as data
from different years), along a specified column. It assumes that the datasets can be
matched by a common column.
}
