% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard-error.R
\name{estimate_with_bootstrap_se}
\alias{estimate_with_bootstrap_se}
\title{Estimate with Bootstrap Standard Errors}
\usage{
estimate_with_bootstrap_se(
  data,
  f,
  weight = "weight",
  repwt_cols = paste0("repwt", 1:4),
  constant = 4/80,
  se_cols,
  ...
)
}
\arguments{
\item{data}{A tibble or data frame containing the data to be analyzed.}

\item{f}{A function that produces new columns or summary statistics.
The function \code{f} must have an argument named \code{weight} to specify the weight column.}

\item{weight}{A string indicating the column name to be used as the main weight.
Defaults to \code{"weight"}.}

\item{repwt_cols}{A vector of strings indicating the names of replicate weight
columns in \code{data}. Defaults to \code{paste0("repwt", 1:4)}.}

\item{constant}{A constant used in the standard error calculation. For IPUMS data,
this is typically \code{4/80}. See \url{https://usa.ipums.org/usa/repwt.shtml} for details.}

\item{se_cols}{A vector of string column names for which standard errors are to be computed.}

\item{...}{Additional arguments passed to the function \code{f}.}
}
\value{
A tibble containing the original main estimate columns along with new
columns for the calculated standard errors. The standard error columns will
have the prefix \code{"se_"}.
}
\description{
Combines bootstrap replicate estimation and standard error calculation into a single function.
It calculates results of a target function \code{f()} using a specified weight column,
computes replicate estimates by substituting each of the specified \code{repwt_cols}
for the \code{weight} argument within \code{f()}, and then calculates standard errors across
specified columns using the output of \code{bootstrap_replicates()}.
}
