% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{extract_factor_label}
\alias{extract_factor_label}
\title{Extract a Column as a Vector from a Lookup Table}
\usage{
extract_factor_label(lookup_table, colname)
}
\arguments{
\item{lookup_table}{A data frame or tibble containing the data.}

\item{colname}{A single character string specifying the name of the column to
extract.}
}
\value{
A vector containing the values from the specified column. If the column
is not found or the inputs are invalid, the function will throw an error.
}
\description{
This function extracts a specified column from a lookup table (data frame or
tibble) and returns it as a vector. It is meant to be used to create factor
labels from data, so that graphs and charts displaying categorical data are
ordered in a logical manner.
}
\details{
The function checks whether the provided \code{lookup_table} is a data frame or tibble,
whether \code{colname} is a single character string, and whether \code{colname} exists within
\code{lookup_table}. If these checks pass, it extracts the column as a vector using \code{dplyr::pull}.
}
