% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucketing.R
\name{join_columns}
\alias{join_columns}
\title{Join Two Datasets on an ID Column with Prioritization}
\usage{
join_columns(data1, data2, column, id)
}
\arguments{
\item{data1}{The deprioritized dataset.}

\item{data2}{The prioritized dataset.}

\item{column}{The name of the column to prioritize and join on.}

\item{id}{The name of the column uniquely identifying observations.}
}
\value{
A dataframe with the joined data and the \code{column} prioritized from \code{data2}.
}
\description{
This function joins two datasets by a common ID column. For a specified column, it prioritizes
values from one dataset over the other. It removes duplicate columns from the output.
}
