% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{map_data}
\alias{map_data}
\title{Create a Colored Map of Shapefiles with Gradient}
\usage{
map_data(data, borders = TRUE)
}
\arguments{
\item{data}{A \code{sf} object (simple feature) containing the shapefile data to be plotted. It must contain
a column called \code{diff} to represent the gradient.}

\item{borders}{A logical indicating whether to include black borders around the shapes. Defaults to \code{TRUE}.}
}
\value{
A \code{ggplot2} plot object displaying a colored map of the shapefile with gradient fill based on the \code{diff} column.
}
\description{
This function generates a colored map using shapefiles, where the color indicates the gradient of the data.
The data must contain a column named \code{diff} for the function to work as intended. Optionally, the function
can add borders to the shapes.
}
\examples{
# Example usage:
# Assuming `sf_data` is a simple feature (sf) object with a `diff` column
map_data(sf_data, borders = TRUE)

}
