% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{map_geographies}
\alias{map_geographies}
\title{Create a Black and White Map of Shapefiles}
\usage{
map_geographies(data)
}
\arguments{
\item{data}{A \code{sf} object (simple feature) containing the shapefile data to be plotted.}
}
\value{
A \code{ggplot2} plot object displaying a black and white map of the shapefile.
}
\description{
This function generates a simple black and white map using shapefiles. It is useful for visualizing
the number of CPUMAs or other geographic boundaries in a region. The function cannot graph data other
than simple shapefiles.
}
\examples{
# Example usage:
# Assuming `sf_data` is a simple feature (sf) object
map_geographies(sf_data)

}
