% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucketing.R
\name{race_eth_bucket}
\alias{race_eth_bucket}
\title{Create a Joint Race/Ethnicity Bucket}
\usage{
race_eth_bucket(data)
}
\arguments{
\item{data}{A dataframe or database table with \code{RACE_bucket} and \code{HISPAN_bucket} columns.}
}
\value{
A dataframe or database table with an additional \code{RACE_ETH_bucket} column.
}
\description{
This function combines the \code{RACE_bucket} and \code{HISPAN_bucket} columns to generate a new joint
\code{RACE_ETH_bucket} column.
}
