% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucketing.R
\name{range_match_lookup}
\alias{range_match_lookup}
\title{Bucket Data Based on Ranges in a Lookup Table}
\usage{
range_match_lookup(data, lookup, input_column, output_column = NULL)
}
\arguments{
\item{data}{A dataframe, tibble, or db object containing the input data to be bucketed.}

\item{lookup}{A dataframe, tibble, or db object containing the lookup table with bucket ranges.}

\item{input_column}{The name of the column in \code{data} to be bucketed.}

\item{output_column}{The name of the output column. Default is \verb{\{input_column\}_bucket}.}
}
\value{
A dataframe, tibble, or db object with an additional column for the assigned bucket.
}
\description{
This function creates custom buckets for continuous variables based on a range-based lookup table.
It appends a new column to the input data with the assigned bucket for each row.
The lookup table must have columns: \code{bucket_name}, \code{lower_bound}, and \code{upper_bound}.
}
