% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard-error.R
\name{se_from_bootstrap}
\alias{se_from_bootstrap}
\title{Calculate Standard Errors Using Bootstrapped Replicate Results}
\usage{
se_from_bootstrap(bootstrap, constant = 4/80, se_cols)
}
\arguments{
\item{bootstrap}{The output from the \code{bootstrap_replicates()} function,
containing the main estimate and replicate estimates.}

\item{constant}{A constant used in the standard error calculation. For IPUMS data,
this is typically \code{4/80}. See \url{https://usa.ipums.org/usa/repwt.shtml} for details.}

\item{se_cols}{A vector of string column names for which standard errors are to be computed.}
}
\value{
A tibble containing the original main estimate columns along with new
columns for the calculated standard errors. The standard error columns will
have the prefix \code{"se_"}.
}
\description{
Computes standard errors across specified columns using the output of
\code{bootstrap_replicates()}. The function assumes that all input tibbles
from bootstrap replicates are ordered identically. If they are not,
(which might happen after a duckdb operation) this will introduce issues.
}
