% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucketing.R
\name{split_lookup_table}
\alias{split_lookup_table}
\title{Split a Lookup Table into Range-Based and Value-Based Components}
\usage{
split_lookup_table(filepath)
}
\arguments{
\item{filepath}{The path to the CSV file containing the lookup table.}
}
\value{
A list with two tibbles: \verb{$value} for value-based lookups, and \verb{$range} for range-based lookups.
}
\description{
This function reads a lookup table from a CSV file and splits it into two components: one for
value-based lookups and one for range-based lookups.
}
