% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucketing.R
\name{write_race_eth_sql_query}
\alias{write_race_eth_sql_query}
\title{Generate SQL Query for Joint Race/Ethnicity Bucket}
\usage{
write_race_eth_sql_query(table)
}
\arguments{
\item{table}{A string. The name of the SQL table where the new column will be added.}
}
\value{
A string. A SQL query with \verb{CASE WHEN} statements to generate the \code{RACE_ETH_bucket}.
}
\description{
This function generates a SQL query to add a \code{RACE_ETH_bucket} column to the database table.
}
