% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucketing.R
\name{write_sql_fragment_range}
\alias{write_sql_fragment_range}
\title{Translate a single line of bucketing logic from a range lookup table into SQL code}
\usage{
write_sql_fragment_range(i, lookup_table, col)
}
\arguments{
\item{i}{Integer. The row index of the lookup table.}

\item{lookup_table}{A tibble or data frame with at least three columns: \code{bucket_name},
\code{lower_bound}, and \code{upper_bound}, which define the range bucketing logic.}

\item{col}{A string. The name of the column in the database that the SQL logic applies to.}
}
\value{
A string. A single line of SQL \verb{CASE WHEN} code for a range-based condition.
}
\description{
This function translates one line of a range-based lookup table into a corresponding
SQL \verb{CASE WHEN} statement.
}
