% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{write_sql_fragment_value}
\alias{write_sql_fragment_value}
\title{Translate a single line of bucketing logic from a value lookup table into SQL code}
\usage{
write_sql_fragment_value(i, lookup_table, col)
}
\arguments{
\item{i}{Integer. The row index of the lookup table.}

\item{lookup_table}{A tibble or data frame with at least two columns: \code{bucket_name}
and \code{specific_value}, which define the bucketing logic.}

\item{col}{A string. The name of the column in the database that the SQL logic applies to.}
}
\value{
A string. A single line of SQL \verb{CASE WHEN} code for a value-based condition.
}
\description{
This function translates one line of a value-based lookup table into a corresponding
SQL \verb{CASE WHEN} statement.
}
