% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucketing.R
\name{write_sql_query}
\alias{write_sql_query}
\title{Create SQL statements to add and populate a new column with bucketing logic}
\usage{
write_sql_query(range_lookup_table, value_lookup_table, col, table)
}
\arguments{
\item{range_lookup_table}{A tibble or data frame with at least three columns:
\code{bucket_name}, \code{lower_bound}, and \code{upper_bound} to define the range bucketing logic.}

\item{value_lookup_table}{A tibble or data frame with at least two columns:
\code{bucket_name} and \code{specific_value} to define the value-based bucketing logic.}

\item{col}{A string. The name of the column in the database that the SQL logic applies to.}

\item{table}{A string. The name of the SQL table where the new column will be added.}
}
\value{
A string containing the SQL statements to add and update the new column.
}
\description{
This function generates SQL \verb{ALTER TABLE} and \code{UPDATE} statements to add a new column
and populate it based on value-based and range-based bucketing rules provided by
\code{value_lookup_table} and \code{range_lookup_table}.
}
