% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{write_sql_query}
\alias{write_sql_query}
\title{Create a full SQL query incorporating both value-based and range-based logic}
\usage{
write_sql_query(range_lookup_table, value_lookup_table, col, table)
}
\arguments{
\item{range_lookup_table}{A tibble or data frame with at least three columns:
\code{bucket_name}, \code{lower_bound}, and \code{upper_bound} to define the range bucketing logic.}

\item{value_lookup_table}{A tibble or data frame with at least two columns:
\code{bucket_name} and \code{specific_value} to define the value-based bucketing logic.}

\item{col}{A string. The name of the column in the database that the SQL logic applies to.}

\item{table}{A string. The name of the SQL table where the new column will be added.}
}
\value{
A string. A full SQL query with \verb{CASE WHEN} statements based on the provided bucketing rules.
}
\description{
This function generates a complete SQL \verb{ALTER TABLE} query that creates a new column
based on value-based and range-based bucketing rules. The bucketing logic is provided
by the \code{value_lookup_table} and \code{range_lookup_table} parameters.
}
